Refactor icalendar helpers into their own module
