+++ khal/khalendar/event.py
@@ -31,9 +31,9 @@ import pytz
 from click import style
 
 from ..exceptions import FatalError
+from ..icalendar import cal_from_ics, delete_instance, invalid_timezone
 from ..terminal import get_color
-from ..utils import (cal_from_ics, delete_instance, generate_random_uid,
-                     invalid_timezone, is_aware, to_naive_utc, to_unix_time)
+from ..utils import generate_random_uid, is_aware, to_naive_utc, to_unix_time
 from ..parse_datetime import timedelta2str
 
 logger = logging.getLogger('khal')
